// restitution-app/ClientApp/src/model/createRestitutionCaseRequestDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Restitution API
 * API for the Restitution Application
 * OpenAPI spec version: v1
 */
import type { ApplicationDto } from './applicationDto';
import type { CourtInfoDto } from './courtInfoDto';
import type { DocumentDto } from './documentDto';
import type { ParticipantDto } from './participantDto';

export interface CreateRestitutionCaseRequestDto {
  application?: ApplicationDto;
  /** @nullable */
  courtInfoCollection?: CourtInfoDto[] | null;
  /** @nullable */
  providerCollection?: ParticipantDto[] | null;
  /** @nullable */
  documentCollection?: DocumentDto[] | null;
}