// restitution-app/ClientApp/src/model/participantDto.ts
/**
 * Generated by orval v8.5.0 🍺
 * Do not edit manually.
 * Restitution API
 * API for the Restitution Application
 * OpenAPI spec version: v1
 */

export interface ParticipantDto {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  middleName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  preferredName?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  addressLine1?: string | null;
  /** @nullable */
  addressLine2?: string | null;
  /** @nullable */
  addressLine3?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  postalCode?: string | null;
  /** @nullable */
  preferredMethodOfContact?: number | null;
  /** @nullable */
  restContactPreferenceForUpdates?: number | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  alternatePhoneNumber?: string | null;
  /** @nullable */
  voicemailOptions?: number | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  custodyLocation?: string | null;
  /** @nullable */
  programName?: string | null;
  /** @nullable */
  relationship1?: string | null;
  /** @nullable */
  relationship2?: string | null;
  /** @nullable */
  relationship2Other?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  contactTitle?: string | null;
  /** @nullable */
  smsPreferred?: number | null;
  /** @nullable */
  isPrimaryEntityContact?: number | null;
}